more changes to ci
